+++ packages/contract-helpers/src/staking-contract/index.ts
@@ -4,6 +4,7 @@ import BaseService from '../commons/BaseService';
 import {
   eEthereumTxType,
   EthereumTransactionTypeExtended,
+  ProtocolAction,
   tEthereumAddress,
   transactionType,
 } from '../commons/types';
@@ -21,8 +22,8 @@ import { ERC20_2612Interface, ERC20_2612Service } from '../erc20-2612';
 import { ERC20Service, IERC20ServiceInterface } from '../erc20-contract';
 import { IAaveStakingHelper } from './typechain/IAaveStakingHelper';
 import { IAaveStakingHelper__factory } from './typechain/IAaveStakingHelper__factory';
-import { IStakedToken } from './typechain/IStakedToken';
-import { IStakedToken__factory } from './typechain/IStakedToken__factory';
+import { IStakedAaveV3 } from './typechain/IStakedAaveV3';
+import { IStakedAaveV3__factory } from './typechain/IStakedAaveV3__factory';
 
 export interface StakingInterface {
   stakingContractAddress: tEthereumAddress;
@@ -51,6 +52,10 @@ export interface StakingInterface {
     amount: string,
     signature: string,
   ) => Promise<EthereumTransactionTypeExtended[]>;
+  claimRewardsAndStake: (
+    user: tEthereumAddress,
+    amount: string,
+  ) => Promise<EthereumTransactionTypeExtended[]>;
 }
 
 type StakingServiceConfig = {
@@ -59,7 +64,7 @@ type StakingServiceConfig = {
 };
 
 export class StakingService
-  extends BaseService<IStakedToken>
+  extends BaseService<IStakedAaveV3>
   implements StakingInterface
 {
   readonly stakingHelperContract: IAaveStakingHelper;
@@ -76,7 +81,7 @@ export class StakingService
     provider: providers.Provider,
     stakingServiceConfig: StakingServiceConfig,
   ) {
-    super(provider, IStakedToken__factory);
+    super(provider, IStakedAaveV3__factory);
 
     this.erc20Service = new ERC20Service(provider);
 
@@ -100,7 +105,7 @@ export class StakingService
     @isPositiveAmount() amount: string,
   ): Promise<string> {
     const { getTokenData } = this.erc20Service;
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
     // eslint-disable-next-line new-cap
@@ -161,7 +166,7 @@ export class StakingService
   ): Promise<EthereumTransactionTypeExtended[]> {
     const txs: EthereumTransactionTypeExtended[] = [];
     const { decimalsOf } = this.erc20Service;
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
     // eslint-disable-next-line new-cap
@@ -199,7 +204,7 @@ export class StakingService
   ): Promise<EthereumTransactionTypeExtended[]> {
     const txs: EthereumTransactionTypeExtended[] = [];
     const { decimalsOf, isApproved, approve } = this.erc20Service;
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
     // eslint-disable-next-line new-cap
@@ -222,7 +227,6 @@ export class StakingService
       txs.push(approveTx);
     }
 
-    console.log(stakingContract);
     const txCallback: () => Promise<transactionType> = this.generateTxCallback({
       rawTxMethod: async () =>
         stakingContract.populateTransaction.stake(
@@ -230,12 +234,17 @@ export class StakingService
           convertedAmount,
         ),
       from: user,
+      action: ProtocolAction.stake,
     });
 
     txs.push({
       tx: txCallback,
       txType: eEthereumTxType.STAKE_ACTION,
-      gas: this.generateTxPriceEstimation(txs, txCallback),
+      gas: this.generateTxPriceEstimation(
+        txs,
+        txCallback,
+        ProtocolAction.stake,
+      ),
     });
 
     return txs;
@@ -247,7 +256,7 @@ export class StakingService
     @isPositiveOrMinusOneAmount() amount: string,
   ): Promise<EthereumTransactionTypeExtended[]> {
     let convertedAmount: string;
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
     if (amount === '-1') {
@@ -281,7 +290,7 @@ export class StakingService
   public cooldown(
     @isEthAddress() user: tEthereumAddress,
   ): EthereumTransactionTypeExtended[] {
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
 
@@ -305,7 +314,7 @@ export class StakingService
     @isPositiveOrMinusOneAmount() amount: string,
   ): Promise<EthereumTransactionTypeExtended[]> {
     let convertedAmount: string;
-    const stakingContract: IStakedToken = this.getContractInstance(
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
       this.stakingContractAddress,
     );
     if (amount === '-1') {
@@ -323,13 +332,61 @@ export class StakingService
         stakingContract.populateTransaction.claimRewards(user, convertedAmount),
       from: user,
       gasSurplus: 20,
+      action: ProtocolAction.claimRewards,
     });
 
     return [
       {
         tx: txCallback,
         txType: eEthereumTxType.STAKE_ACTION,
-        gas: this.generateTxPriceEstimation([], txCallback),
+        gas: this.generateTxPriceEstimation(
+          [],
+          txCallback,
+          ProtocolAction.claimRewards,
+        ),
+      },
+    ];
+  }
+
+  @StakingValidator
+  public async claimRewardsAndStake(
+    @isEthAddress() user: tEthereumAddress,
+    @isPositiveOrMinusOneAmount() amount: string,
+  ): Promise<EthereumTransactionTypeExtended[]> {
+    let convertedAmount: string;
+    const stakingContract: IStakedAaveV3 = this.getContractInstance(
+      this.stakingContractAddress,
+    );
+    if (amount === '-1') {
+      convertedAmount = constants.MaxUint256.toString();
+    } else {
+      const { decimalsOf } = this.erc20Service;
+      // eslint-disable-next-line new-cap
+      const stakedToken: string = await stakingContract.STAKED_TOKEN();
+      const stakedTokenDecimals: number = await decimalsOf(stakedToken);
+      convertedAmount = valueToWei(amount, stakedTokenDecimals);
+    }
+
+    const txCallback: () => Promise<transactionType> = this.generateTxCallback({
+      rawTxMethod: async () =>
+        stakingContract.populateTransaction.claimRewardsAndStake(
+          user,
+          convertedAmount,
+        ),
+      from: user,
+      gasSurplus: 20,
+      action: ProtocolAction.claimRewardsAndStake,
+    });
+
+    return [
+      {
+        tx: txCallback,
+        txType: eEthereumTxType.STAKE_ACTION,
+        gas: this.generateTxPriceEstimation(
+          [],
+          txCallback,
+          ProtocolAction.claimRewardsAndStake,
+        ),
       },
     ];
   }