+++ packages/contract-helpers/src/uiStakeDataProvider-contract/_mocks.ts
@@ -1,8 +1,12 @@
 import { BigNumber } from 'ethers';
 
 export const GeneralStakeUIDataRaw = {
-  0: {
-    stakeTokenTotalSupply: BigNumber.from({
+  stkAaveData: {
+    stakedTokenTotalSupply: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    stakedTokenTotalRedeemableAmount: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -14,7 +18,7 @@ export const GeneralStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
-    stakeTokenPriceEth: BigNumber.from({
+    stakedTokenPriceEth: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -66,9 +70,17 @@ export const GeneralStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    8: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
   },
-  1: {
-    stakeTokenTotalSupply: BigNumber.from({
+  stkBptData: {
+    stakedTokenTotalSupply: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    stakedTokenTotalRedeemableAmount: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -80,7 +92,7 @@ export const GeneralStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
-    stakeTokenPriceEth: BigNumber.from({
+    stakedTokenPriceEth: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -132,16 +144,21 @@ export const GeneralStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    8: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
   },
-  2: BigNumber.from({
+  ethPrice: BigNumber.from({
     _hex: '0x0',
     _isBigNumber: true,
   }),
 };
 
 export const GetUserStakeUIDataRaw = {
-  0: {
-    stakeTokenUserBalance: BigNumber.from({
+  ...GeneralStakeUIDataRaw,
+  stkAaveUserData: {
+    stakedTokenUserBalance: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -149,15 +166,16 @@ export const GetUserStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
-    userCooldown: BigNumber.from({
+    stakedTokenRedeemableAmount: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    userCooldownAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCoolddownTimestamp: 0,
     userIncentivesToClaim: BigNumber.from({
-      _hex: '0x0',
-      _isBigNumber: true,
-    }),
-    userPermitNonce: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -181,9 +199,14 @@ export const GetUserStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    5: 0,
+    6: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
   },
-  1: {
-    stakeTokenUserBalance: BigNumber.from({
+  stkBptUserData: {
+    stakedTokenUserBalance: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -191,15 +214,16 @@ export const GetUserStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
-    userCooldown: BigNumber.from({
+    stakedTokenRedeemableAmount: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    userCooldownAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCooldownTimestamp: 0,
     userIncentivesToClaim: BigNumber.from({
-      _hex: '0x0',
-      _isBigNumber: true,
-    }),
-    userPermitNonce: BigNumber.from({
       _hex: '0x0',
       _isBigNumber: true,
     }),
@@ -223,8 +247,141 @@ export const GetUserStakeUIDataRaw = {
       _hex: '0x0',
       _isBigNumber: true,
     }),
+    5: 0,
+    6: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
   },
-  2: BigNumber.from({
+  ethPrice: BigNumber.from({
+    _hex: '0x0',
+    _isBigNumber: true,
+  }),
+};
+
+export const GetUserStakeTotalsRaw = {
+  ...GeneralStakeUIDataRaw,
+  stkAaveUserData: {
+    stakedTokenUserBalance: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    underlyingTokenUserBalance: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    stakedTokenRedeemableAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCooldownAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userIncentivesToClaim: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCooldownTimestamp: 0,
+    rewardsToClaim: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    0: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    1: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    2: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    3: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    4: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    5: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    6: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    7: 0,
+    8: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+  },
+  stkBptUserData: {
+    stakedTokenUserBalance: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    underlyingTokenUserBalance: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    stakedTokenRedeemableAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCooldownAmount: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userIncentivesToClaim: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    userCooldownTimestamp: 0,
+    rewardsToClaim: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    0: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    1: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    2: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    3: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    4: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    5: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    6: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+    7: 0,
+    8: BigNumber.from({
+      _hex: '0x0',
+      _isBigNumber: true,
+    }),
+  },
+  ethPrice: BigNumber.from({
     _hex: '0x0',
     _isBigNumber: true,
   }),
@@ -233,6 +390,7 @@ export const GetUserStakeUIDataRaw = {
 export const GeneralStakeUIDataHumanized = {
   aave: {
     stakeTokenTotalSupply: '0',
+    stakeTokenTotalRedeemableAmount: '0',
     stakeCooldownSeconds: 0,
     stakeUnstakeWindow: 0,
     stakeTokenPriceEth: '0',
@@ -243,6 +401,7 @@ export const GeneralStakeUIDataHumanized = {
   },
   bpt: {
     stakeTokenTotalSupply: '0',
+    stakeTokenTotalRedeemableAmount: '0',
     stakeCooldownSeconds: 0,
     stakeUnstakeWindow: 0,
     stakeTokenPriceEth: '0',
@@ -251,23 +410,25 @@ export const GeneralStakeUIDataHumanized = {
     distributionPerSecond: '0',
     distributionEnd: '0',
   },
-  usdPriceEth: '0',
+  ethPriceUsd: '0',
 };
 
 export const GetUserStakeUIDataHumanized = {
   aave: {
     stakeTokenUserBalance: '0',
     underlyingTokenUserBalance: '0',
-    userCooldown: 0,
+    stakeTokenRedeemableAmount: '0',
+    userCooldownTimestamp: 0,
+    userCooldownAmount: '0',
     userIncentivesToClaim: '0',
-    userPermitNonce: '0',
   },
   bpt: {
     stakeTokenUserBalance: '0',
     underlyingTokenUserBalance: '0',
-    userCooldown: 0,
+    stakeTokenRedeemableAmount: '0',
+    userCooldownTimestamp: 0,
+    userCooldownAmount: '0',
     userIncentivesToClaim: '0',
-    userPermitNonce: '0',
   },
-  usdPriceEth: '0',
+  ethPriceUsd: '0',
 };